third_party/blink/web_tests/svg/dynamic-updates/SVGFEComponentTransferElement-svgdom-type-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'type' property of the SVGFEComponentTransferElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS feRFunc.type.baseVal is SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_GAMMA
PASS feGFunc.type.baseVal is SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_GAMMA
PASS feBFunc.type.baseVal is SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_GAMMA
PASS feAFunc.type.baseVal is SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_GAMMA
PASS feRFunc.type.baseVal is SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_LINEAR
PASS feGFunc.type.baseVal is SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_LINEAR
PASS feBFunc.type.baseVal is SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_LINEAR
PASS feAFunc.type.baseVal is SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_LINEAR
PASS successfullyParsed is true

TEST COMPLETE
